[Rudra Raj] : Created a program to calculate volume of cylinder.
corrected the data type.

+++ gcr-codebase/programming-prequisites/CylinderVolume.java
@@ -4,10 +4,10 @@ import java.util.Scanner;
      public static void main(String[] args){
 	    
 		Scanner sc = new Scanner(System.in);
-		float r = sc.nextFloat();
-		float h = sc.nextFloat();
+		double r = sc.nextDouble();
+		double h = sc.nextDouble();
 		
-		float volume = 3.14 * r*r * h;
+		double volume = 3.14 * r*r * h;
 		
 		System.out.println(volume);
 	}
